Refactor: Remove mapping exceptions inside AuthenticationUseCase and repositories and usecases.
