+++ .idea/gradle.xml
@@ -16,5 +16,6 @@
         </option>
       </GradleProjectSettings>
     </option>
+    <option name="offlineMode" value="true" />
   </component>
 </project>
