+++ app/src/main/java/com/example/cyclistance/feature_authentication/data/repository/AuthRepositoryImpl.kt
@@ -1,24 +1,13 @@
 package com.example.cyclistance.feature_authentication.data.repository
 
 import android.content.Context
-import android.net.Uri
-import android.os.Bundle
 import com.example.cyclistance.R
 import com.example.cyclistance.core.utils.connection.ConnectionStatus.hasInternetConnection
-import com.example.cyclistance.core.utils.constants.AuthConstants.DATA_STORE_PHONE_NUMBER_KEY
 import com.example.cyclistance.core.utils.constants.AuthConstants.FACEBOOK_CONNECTION_FAILURE
-import com.example.cyclistance.core.utils.constants.AuthConstants.IMAGE_LARGE_SIZE
-import com.example.cyclistance.core.utils.constants.AuthConstants.IMAGE_SMALL_SIZE
 import com.example.cyclistance.core.utils.constants.AuthConstants.USER_NOT_FOUND
-import com.example.cyclistance.core.utils.extension.editData
-import com.example.cyclistance.core.utils.extension.getData
 import com.example.cyclistance.feature_authentication.domain.exceptions.AuthExceptions
 import com.example.cyclistance.feature_authentication.domain.model.SignInCredential
 import com.example.cyclistance.feature_authentication.domain.repository.AuthRepository
-import com.example.cyclistance.feature_mapping.data.repository.dataStore
-import com.facebook.AccessToken
-import com.facebook.GraphRequest
-import com.google.firebase.FirebaseException
 import com.google.firebase.FirebaseNetworkException
 import com.google.firebase.FirebaseTooManyRequestsException
 import com.google.firebase.auth.FacebookAuthProvider
@@ -27,11 +16,7 @@ import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException
 import com.google.firebase.auth.FirebaseAuthInvalidUserException
 import com.google.firebase.auth.FirebaseAuthUserCollisionException
 import com.google.firebase.auth.GoogleAuthProvider
-import com.google.firebase.auth.ktx.userProfileChangeRequest
-import com.google.firebase.storage.FirebaseStorage
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.suspendCancellableCoroutine
-import org.json.JSONObject
 import timber.log.Timber
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
@@ -40,33 +25,10 @@ import kotlin.coroutines.resumeWithException
 class AuthRepositoryImpl(
     private val context: Context,
     private val auth: FirebaseAuth,
-    private val storage: FirebaseStorage
     ) : AuthRepository {
 
-    private var dataStore = context.dataStore
 
-    override suspend fun uploadImage(v: String): String {
-        val reference = storage.reference.child("images/${getId()}")
-        val uploadTask = reference.putFile(Uri.parse(v))
-        return suspendCancellableCoroutine { continuation ->
-            uploadTask.addOnCompleteListener { task ->
-                task.exception?.let { exception ->
-                    if (exception is FirebaseNetworkException) {
-                        continuation.resumeWithException(
-                            AuthExceptions.NetworkException(
-                                message = context.getString(R.string.no_internet_message)))
-                    }
-                }
 
-                if (task.isSuccessful) {
-                    reference.downloadUrl.addOnSuccessListener{
-                        continuation.resume(it.toString())
-                    }
-                }
-
-            }
-        }
-    }
 
     override suspend fun reloadEmail(): Boolean {
         return suspendCancellableCoroutine { continuation ->
@@ -245,36 +207,7 @@ class AuthRepositoryImpl(
 
 
 
-    private suspend fun getUserFacebookInformation(): String{
-        return suspendCancellableCoroutine { continuation ->
-            val request = GraphRequest.newMeRequest(getFacebookToken()) { jsonObject: JSONObject?, _ ->
-                continuation.resume(jsonObject.toString())
-            }
-            val parameters = Bundle()
-            parameters.putString("fields", "id,name, email, link, picture.type(large)")
-            request.parameters = parameters
-            request.executeAsync()
-        }
-    }
 
-    private fun getFacebookToken(): AccessToken?{
-        return AccessToken.getCurrentAccessToken()
-    }
-
-    override suspend fun getName(): String? {
-        val infoString = getUserFacebookInformation()
-        val jsonObject = JSONObject(infoString)
-        val fbName = jsonObject.getString("name")
-        return auth.currentUser?.displayName?.takeIf { it.isNotEmpty() } ?: fbName
-    }
-
-    override suspend fun getPhotoUrl(): String {
-        val infoString = getUserFacebookInformation()
-        val jsonObject = JSONObject(infoString)
-        val fbPhotoUrl = jsonObject.getJSONObject("picture").getJSONObject("data").getString("url")
-        return auth.currentUser?.photoUrl.toString().replace(oldValue = IMAGE_SMALL_SIZE, newValue = IMAGE_LARGE_SIZE).takeIf { it.isNotEmpty() } ?: fbPhotoUrl
-
-    }
 
     override fun isSignedInWithProvider(): Boolean? {
         return auth.currentUser?.providerData?.any {
@@ -292,50 +225,7 @@ class AuthRepositoryImpl(
     }
 
 
-    override suspend fun updateProfile(photoUrl: String?, name: String?): Boolean {
-        val profileUpdates = userProfileChangeRequest {
-            name?.let { this.displayName = it }
-            photoUrl?.let { this.photoUri = Uri.parse(photoUrl) }
-        }
 
 
-        if(!context.hasInternetConnection()){
-            throw AuthExceptions.NetworkException(message = context.getString(R.string.no_internet_message))
-        }
-
-        return suspendCancellableCoroutine { continuation ->
-            auth.currentUser?.updateProfile(profileUpdates)
-                ?.addOnCompleteListener { updateProfile ->
-                    updateProfile.exception?.let { exception ->
-
-                        if (exception is FirebaseNetworkException) {
-                            continuation.resumeWithException(
-                                AuthExceptions.NetworkException(
-                                    message = context.getString(R.string.no_internet_message)))
-                        }
-
-                        if(exception is FirebaseException){
-
-                            continuation.resumeWithException(
-                                AuthExceptions.InternalServerException(
-                                    message = exception.message ?: context.getString(R.string.somethingWentWrong))
-                                )
-                        }
-
-                    }
-                    if (continuation.isActive) {
-                        continuation.resume(updateProfile.isSuccessful)
-                    }
-                }
-        }
-    }
-
-    override suspend fun updatePhoneNumber(phoneNumber: String) {
-        dataStore.editData(DATA_STORE_PHONE_NUMBER_KEY, phoneNumber)
-    }
-
-    override fun getPhoneNumber(): Flow<String> {
-        return dataStore.getData(key = DATA_STORE_PHONE_NUMBER_KEY, defaultValue = "")
-    }
 
 }
