+++ app/src/main/java/com/example/cyclistance/feature_authentication/domain/exceptions/AuthExceptions.kt
@@ -9,5 +9,5 @@ sealed class AuthExceptions {
     class PasswordException(message: String) : RuntimeException(message)
     class ConfirmPasswordException(message: String) : RuntimeException(message)
     class EmailException(message: String) : RuntimeException(message)
-    class InternalServerException(message: String) : RuntimeException(message)
+    class UserException(message: String = "User not found"):RuntimeException(message)
 }
