+++ app/src/main/java/com/example/cyclistance/feature_authentication/domain/use_case/AuthenticationUseCase.kt
@@ -1,27 +1,28 @@
 package com.example.cyclistance.feature_authentication.domain.use_case
 
-import com.example.cyclistance.feature_authentication.domain.use_case.create_account.*
-import com.example.cyclistance.feature_authentication.domain.use_case.read_account.*
+import com.example.cyclistance.feature_authentication.domain.use_case.create_account.CreateWithEmailAndPasswordUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.create_account.SignInWithCredentialUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.create_account.SignInWithEmailAndPasswordUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.read_account.GetEmailUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.read_account.GetIdUseCase
 import com.example.cyclistance.feature_authentication.domain.use_case.sign_out_account.SignOutUseCase
-import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.*
+import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.HasAccountSignedInUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.IsEmailVerifiedUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.IsSignedInWithProviderUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.ReloadEmailUseCase
+import com.example.cyclistance.feature_authentication.domain.use_case.verify_account.SendEmailVerificationUseCase
 
 data class AuthenticationUseCase(
     val reloadEmailUseCase: ReloadEmailUseCase,
     val signOutUseCase: SignOutUseCase,
     val createWithEmailAndPasswordUseCase: CreateWithEmailAndPasswordUseCase,
     val getEmailUseCase: GetEmailUseCase,
-    val getNameUseCase: GetNameUseCase,
     val getIdUseCase: GetIdUseCase,
-    val getPhoneNumberUseCase: GetPhoneNumberUseCase,
-    val getPhotoUrlUseCase: GetPhotoUrlUseCase,
     val hasAccountSignedInUseCase: HasAccountSignedInUseCase,
     val isEmailVerifiedUseCase: IsEmailVerifiedUseCase,
     val isSignedInWithProviderUseCase: IsSignedInWithProviderUseCase,
     val sendEmailVerificationUseCase: SendEmailVerificationUseCase,
     val signInWithEmailAndPasswordUseCase: SignInWithEmailAndPasswordUseCase,
     val signInWithCredentialUseCase: SignInWithCredentialUseCase,
-    val updatePhoneNumberUseCase : UpdatePhoneNumberUseCase,
-    val updateProfileUseCase : UpdateProfileUseCase,
-    val uploadImageUseCase: UploadImageUseCase
 
     )
