+++ app/src/main/java/com/example/cyclistance/feature_settings/domain/repository/SettingRepository.kt
@@ -6,4 +6,10 @@ interface SettingRepository {
 
     suspend fun toggleTheme(value: Boolean)
     fun isDarkTheme(): Flow<Boolean>
+    suspend fun updatePhoneNumber(phoneNumber: String)
+    fun getPhoneNumber(): Flow<String>
+    suspend fun updateProfile(s: String?, name: String?): Boolean
+    suspend fun uploadImage(v: String): String
+    suspend fun getName(): String?
+    suspend fun getPhotoUrl(): String?
 }
